+++ .env.production
@@ -1,0 +1,2 @@
+GATSBY_SHOP_NAME=hempuptest
+GATSBY_SHOPIFY_ACCESS_TOKEN=1087437fc7ba5e6705705deb2121c02e
