+++ package.json
@@ -16,6 +16,7 @@
     "gatsby-plugin-sharp": "^2.2.7",
     "gatsby-plugin-styled-components": "^3.1.2",
     "gatsby-source-filesystem": "^2.1.5",
+    "gatsby-source-shopify2": "^2.0.7",
     "gatsby-transformer-sharp": "^2.2.3",
     "prop-types": "^15.7.2",
     "react": "^16.8.6",
