+++ src/pages/store/index.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import { Link } from 'gatsby'
+import Layout from '../components/layout'
+import SEO from '../components/seo'
+import Header from '../components/header';
+import '../styles/landing.scss'
+
+const Landing = () => (
+    <Layout>
+        <SEO title='Hemp Up - Store' keywords={['cbd', 'oils', 'edibles']}/>
+        
+    </Layout>
+)
+
+export default Landing
